implement thrift interface

+++ sims/emerysimple/agent_interface.go
@@ -10,8 +10,8 @@ type SpaceSpec struct {
 	// Continuous
 	ContinuousShape []int   `desc:"The dimensions of an array. For example, [3,2] would be a 3 by 2 array. [1] would be a single value."`
 	Stride          []int   // TODO Describe
-	Min             float32 `desc:"The minimum continuous value."`
-	Max             float32 `desc:"The maximum continuous value."`
+	Min             float64 `desc:"The minimum continuous value."`
+	Max             float64 `desc:"The maximum continuous value."`
 
 	// Discrete
 	DiscreteLabels []string `desc:"The names of the discrete possibilities, such as ['left', 'right']. The length of this is the number of discrete possibilities that this shape encapsulates."`
